chore: rename file as per PR comments
