+++ frontend/hooks/prevent-window-resize/index.tsx
