+++ frontend/hooks/prevent-window-resize/prevent-window-resize.spec.tsx
